refactor(homepage): remove LinkList component
